trying to add custom manage.py command

+++ posts/managment/commands/deleteusers.py
@@ -7,4 +7,5 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         all_users = User.objects.all()
         for user in all_users:
-            user.delete()
+            if user.username != "koi":
+                user.delete()
